requirements.txt
requests
dotenv
google-generativeai
pygame
numpy